requirements.txt
fastapi==0.87.0
matplotlib==3.5.1
numpy>=1.17
scipy==1.8.0
uvicorn[standard]
pyyaml~=6.0